pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
version = "3.0.0"
name = "clinical_ETL"
dependencies = [
    "pandas>=2.1.0",
    "pytest>=7.2.0",
    "pyYAML>=5.4.1",
    "dateparser>=1.1.0",
    "openpyxl>=3.0.9",
    "requests>=2.29.0",
    "jsonschema~=4.19.2",
    "openapi-spec-validator>=0.7.1",
    "pdoc3>=0.10.0",
    "tqdm>=4.66.2"
    ]
requires-python = ">= 3.10"
description = "ETL module for transforming clinical CSV data into properly-formatted packets for ingest into Katsu"
readme = "README.md"

[project.scripts]
CSVConvert = "clinical_etl.CSVConvert:main"

[project.urls]
Repository = "https://github.com/CanDIG/clinical_ETL_code"